+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added ForEach extension with the additional index parameter for each item.
+### Supported .Net Versions
+- .Net 8
+- .Net 9
 
 ## [2.2.0] - 2024-12-01
 ### Added